Nkaashoek unginkgo (#13)
Full conversion away from Ginkgo into a custom backend/framework. This should make it much easier in the future to control exactly what is being test and in what order. Already the code is simpler and easier to understand than ginkgo.

This is just the initial commit, more is coming down the pipeline to bring everything inline and up to date.